Add more tests of async iterator usage
